feat: remove extraneous esbuild and jest configs
This commits removes the esbuild and jest configs from any
individual worker directory. This is because the configs are
now in the root of the project and are shared across all workers.

`esbuild` is now the default bundler for all workers, provided by `wrangler`.

`jest` is now the default test runner for all workers, provided by `jest` (using `ts-jest`).